add set and use flags for picard stuff
Found on Github (gentoo/gentoo) PR 4196; comment
===
The dependency on dev-python/python-discid should be removed, since it is an optional runtime dependency and you don't need to rebuild media-sound/picard to enable CD lookup support.

See https://devmanual.gentoo.org/general-concepts/use-flags/

The usage of a USE flag should not control runtime dependencies when the package does not link to it. Doing so will create extra configuration for the package and re-compilation for no underlying file change on disk. This should be avoided and instead can be conveyed to the user via post install messages if needed.===

Signed-off-by: Markus Rathgeb <maggu2810@gmail.com>